Estimate relationship between reps and weight using the non-linear mixed-effects regression

These functions provide estimated 1RM and parameter values using the mixed-effect regression. By default, target variable is the reps performed, while the predictor is the perc_1RM or weight. To reverse this, use the reverse = TRUE argument

estimate_k_mixed(athlete, perc_1RM, reps, eRIR = 0, reverse = FALSE, ...)

estimate_k_1RM_mixed(
  athlete,
  weight,
  reps,
  eRIR = 0,
  reverse = FALSE,
  random = k + oneRM ~ 1,
  ...
)

estimate_kmod_mixed(athlete, perc_1RM, reps, eRIR = 0, reverse = FALSE, ...)

estimate_kmod_1RM_mixed(
  athlete,
  weight,
  reps,
  eRIR = 0,
  reverse = FALSE,
  random = kmod + oneRM ~ 1,
  ...
)

estimate_klin_mixed(athlete, perc_1RM, reps, eRIR = 0, reverse = FALSE, ...)

estimate_klin_1RM_mixed(
  athlete,
  weight,
  reps,
  eRIR = 0,
  reverse = FALSE,
  random = klin + oneRM ~ 1,
  ...
)

Arguments

athlete

Athlete identifier

perc_1RM

%1RM

reps

Number of repetitions done

eRIR

Subjective estimation of reps-in-reserve (eRIR)

reverse

Logical, default is FALSE. Should reps be used as predictor instead as a target?

...

Forwarded to nlme function

weight

Weight used

random

Random parameter forwarded to nlme function. Default is k + zeroRM ~ 1

Value

nlme object

Functions

estimate_k_mixed: Estimate the parameter k in the Epley's equation

estimate_k_1RM_mixed: Estimate the parameter k in the Epley's equation, as well as 1RM. This is a novel estimation function that uses the absolute weights

estimate_kmod_mixed: Estimate the parameter kmod in the Modified Epley's equation

estimate_kmod_1RM_mixed: Estimate the parameter kmod in the Modified Epley's equation, as well as 1RM. This is a novel estimation function that uses the absolute weights

estimate_klin_mixed: Estimate the parameter klin in the Linear/Brzycki's equation

estimate_klin_1RM_mixed: Estimate the parameter klin in the Linear/Brzycki equation, as well as 1RM. This is a novel estimation function that uses the absolute weights

Examples

# ---------------------------------------------------------
# Epley's model
m1 <- estimate_k_mixed(
  athlete = RTF_testing$Athlete,
  perc_1RM = RTF_testing$`Real %1RM`,
  reps = RTF_testing$nRM
)

coef(m1)
#>                    k
#> Athlete A 0.01937865
#> Athlete B 0.03403605
#> Athlete C 0.06747237
#> Athlete D 0.02754989
#> Athlete E 0.04001677
#> Athlete F 0.02442086
#> Athlete G 0.03584091
#> Athlete H 0.02952992
#> Athlete I 0.02172886
#> Athlete J 0.04741195
#> Athlete K 0.05844494
#> Athlete L 0.03987504
# ---------------------------------------------------------
# Epley's model that also estimates 1RM
m1 <- estimate_k_1RM_mixed(
  athlete = RTF_testing$Athlete,
  weight = RTF_testing$`Real Weight`,
  reps = RTF_testing$nRM
)
#> Warning: Iteration 1, LME step: nlminb() did not converge (code = 1). PORT message: false convergence (8)
#> Warning: Iteration 3, LME step: nlminb() did not converge (code = 1). PORT message: false convergence (8)

coef(m1)
#>                    k     oneRM
#> Athlete A 0.02009581  98.92157
#> Athlete B 0.03190781  90.67941
#> Athlete C 0.05146494 107.26821
#> Athlete D 0.02931825 103.66577
#> Athlete E 0.04599428 109.10680
#> Athlete F 0.02446026  87.88554
#> Athlete G 0.03321014  97.45520
#> Athlete H 0.03082763 127.49806
#> Athlete I 0.02224339 105.78176
#> Athlete J 0.04192211  86.29503
#> Athlete K 0.05472413  95.77180
#> Athlete L 0.03826351 133.45715
# ---------------------------------------------------------
# Modifed Epley's model
m1 <- estimate_kmod_mixed(
  athlete = RTF_testing$Athlete,
  perc_1RM = RTF_testing$`Real %1RM`,
  reps = RTF_testing$nRM
)

coef(m1)
#>                 kmod
#> Athlete A 0.02061064
#> Athlete B 0.03816219
#> Athlete C 0.07955734
#> Athlete D 0.03001988
#> Athlete E 0.04559265
#> Athlete F 0.02638650
#> Athlete G 0.04044088
#> Athlete H 0.03240645
#> Athlete I 0.02325264
#> Athlete J 0.05521785
#> Athlete K 0.06915300
#> Athlete L 0.04531026
# ---------------------------------------------------------
# Modified Epley's model that also estimates 1RM
m1 <- estimate_kmod_1RM_mixed(
  athlete = RTF_testing$Athlete,
  weight = RTF_testing$`Real Weight`,
  reps = RTF_testing$nRM
)
#> Warning: Iteration 1, LME step: nlminb() did not converge (code = 1). PORT message: false convergence (8)

coef(m1)
#>                 kmod     oneRM
#> Athlete A 0.03015273 109.90560
#> Athlete B 0.03302734  91.99512
#> Athlete C 0.03135543 102.41451
#> Athlete D 0.03086467 105.47094
#> Athlete E 0.03133881 102.51789
#> Athlete F 0.03261366  94.57202
#> Athlete G 0.03213301  97.56804
#> Athlete H 0.02771236 125.11415
#> Athlete I 0.02956070 113.59505
#> Athlete J 0.03421898  84.57022
#> Athlete K 0.03325611  90.57083
#> Athlete L 0.02747206 126.61250
# ---------------------------------------------------------
# Linear/Brzycki model
m1 <- estimate_klin_mixed(
  athlete = RTF_testing$Athlete,
  perc_1RM = RTF_testing$`Real %1RM`,
  reps = RTF_testing$nRM
)

coef(m1)
#>               klin
#> Athlete A 64.11011
#> Athlete B 35.11002
#> Athlete C 16.22552
#> Athlete D 45.24156
#> Athlete E 30.05447
#> Athlete F 50.70299
#> Athlete G 33.20719
#> Athlete H 41.90200
#> Athlete I 57.35190
#> Athlete J 24.65495
#> Athlete K 19.43184
#> Athlete L 30.18074
# ---------------------------------------------------------
# Linear/Brzycki model that also estimates 1RM
m1 <- estimate_klin_1RM_mixed(
  athlete = RTF_testing$Athlete,
  weight = RTF_testing$`Real Weight`,
  reps = RTF_testing$nRM
)
#> Warning: Iteration 1, LME step: nlminb() did not converge (code = 1). PORT message: false convergence (8)

coef(m1)
#>               klin     oneRM
#> Athlete A 75.15769  96.19463
#> Athlete B 45.44975  88.91455
#> Athlete C 25.33309 107.11406
#> Athlete D 53.32681 100.49170
#> Athlete E 33.89090 106.06520
#> Athlete F 62.45929  85.33938
#> Athlete G 43.19576  95.83903
#> Athlete H 50.16078 123.79762
#> Athlete I 67.24280 103.20196
#> Athlete J 34.37575  84.98052
#> Athlete K 24.84090  95.45042
#> Athlete L 39.44232 130.14431