Generic progression table

Generic progression tables allow for custom progressions using rep_start, rep_step, inc_start, and inc_step parameters. For more information check the 'Strength Training Manual' Chapter 6.

percMR_step_generic(
  reps,
  step = 0,
  rep_start = 0,
  rep_step = 0,
  inc_start = 1,
  inc_step = ((2 - 1)/11),
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM_percMR,
  ...
)

RIR_increment_generic(
  reps,
  step = 0,
  rep_start = 2,
  rep_step = ((6 - 2)/11),
  inc_start = 1,
  inc_step = ((2 - 1)/11),
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM,
  ...
)

perc_drop_generic(
  reps,
  step = 0,
  rep_start = -0.05,
  rep_step = ((-0.1 - -0.05)/11),
  inc_start = -0.025,
  inc_step = ((-0.05 - -0.025)/11),
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM,
  ...
)

relInt_generic(
  reps,
  step = 0,
  rep_start = 0,
  rep_step = 0,
  inc_start = 1,
  inc_step = ((2 - 1)/11),
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM_relInt,
  ...
)

Arguments

reps

Number of reps

step

Progression step. Default is 0. Use negative numbers (i.e., -1, -2)

rep_start

Used to define intensive, normal, and extensive progression. rep_start defines the adjustment for the first rep

rep_step

Used to define intensive, normal, and extensive progression. rep_step defines the adjustment as rep number increases

inc_start

Defines the progression for step for single rep

inc_step

Defines the progression for step for rep increase For example, lower reps can have bigger jumps than high reps.

adjustment

Additional adjustment. Default is 0.

func_max_perc_1RM

What max rep table should be used? Default is get_max_perc_1RM

...

Forwarded to func_max_perc_1RM. Used to differentiate between 'grinding' and 'ballistic' max reps schemes.

Value

List with two elements: adjustment and perc_1RM

Functions

percMR_step_generic: %MR Step generic table

RIR_increment_generic: RIR Increment generic table

perc_drop_generic: Percent Drop generic table

relInt_generic: Relative Intensity generic table

Examples

percMR_step_generic(10, step = seq(-3, 0, 1))
#> $adjustment
#> [1] 6.454545 4.636364 2.818182 1.000000
#> 
#> $perc_1RM
#> [1] 0.9509396 0.9329894 0.8943254 0.7501875
#> 
percMR_step_generic(5, step = seq(-3, 0, 1), type = "ballistic")
#> $adjustment
#> [1] 5.090909 3.727273 2.363636 1.000000
#> 
#> $perc_1RM
#> [1] 0.9386052 0.9179858 0.8765128 0.7501875
#> 
RIR_increment_generic(10, step = seq(-3, 0, 1))
#> $adjustment
#> [1] 10.727273  8.909091  7.090909  5.272727
#> 
#> $perc_1RM
#> [1] 0.5916396 0.6136201 0.6372969 0.6628742
#> 
RIR_increment_generic(5, step = seq(-3, 0, 1), type = "ballistic")
#> $adjustment
#> [1] 7.545455 6.181818 4.818182 3.454545
#> 
#> $perc_1RM
#> [1] 0.5448026 0.5731615 0.6046348 0.6397655
#> 
perc_drop_generic(10, step = seq(-3, 0, 1))
#> $adjustment
#> [1] -0.22727273 -0.18181818 -0.13636364 -0.09090909
#> 
#> $perc_1RM
#> [1] 0.5229148 0.5683694 0.6138239 0.6592785
#> 
perc_drop_generic(5, step = seq(-3, 0, 1), type = "ballistic")
#> $adjustment
#> [1] -0.17045455 -0.13636364 -0.10227273 -0.06818182
#> 
#> $perc_1RM
#> [1] 0.5797330 0.6138239 0.6479148 0.6820057
#> 
relInt_generic(10, step = seq(-3, 0, 1))
#> $adjustment
#> [1] 6.454545 4.636364 2.818182 1.000000
#> 
#> $perc_1RM
#> [1] 4.8421196 3.4781423 2.1141649 0.7501875
#> 
relInt_generic(5, step = seq(-3, 0, 1), type = "ballistic")
#> $adjustment
#> [1] 5.090909 3.727273 2.363636 1.000000
#> 
#> $perc_1RM
#> [1] 3.8191366 2.7961536 1.7731706 0.7501875
#>