Max-Reps Tables

Family of functions to represent max reps tables

get_max_reps(perc_1RM, RIR = 0, type = "grinding")

get_max_perc_1RM(max_reps, RIR = 0, type = "grinding")

get_predicted_1RM(weight, reps, RIR = 0, type = "grinding")

Arguments

perc_1RM
Percent of 1RM

RIR
Reps In Reserve

type
Type of max rep table. Options are grinding (Default) and ballistic.

max_reps
Maximum repetitions

weight
Weight used

reps
Number of repetitions done

Functions

get_max_reps: Get maximum number of repetitions

get_max_perc_1RM: Get maximum %1RM

get_predicted_1RM: Get predicted 1RM

Examples

# Get max reps that can be done with 75% 1RM
get_max_reps(0.75)

#> [1] 10.01

# Get max reps that can be done with 80% 1RM with 2 reps in reserve
get_max_reps(
  perc_1RM = 0.8,
  RIR = 3
)

#> [1] 4.5075

# Get max reps that can be done with 90% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps(
  perc_1RM = 0.9,
  RIR = 2,
  type = "ballistic"
)

#> [1] -0.3316667

# Get max %1RM to be used when doing 5 reps to failure
get_max_perc_1RM(5)

#> [1] 0.8572653

# Get max %1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM(
  max_reps = 3,
  RIR = 2
)

#> [1] 0.8572653

# Get max %1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM(
  max_reps = 3,
  RIR = 2,
  type = "ballistic"
)

#> [1] 0.7501875

# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM(
  weight = 100,
  reps = 5
)

#> [1] 116.65

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM(
  weight = 120,
  reps = 3,
  RIR = 2
)

#> [1] 139.98

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM(
  weight = 120,
  reps = 2,
  RIR = 1,
  type = "ballistic"
)

#> [1] 143.976