Repetition maximum testing of 12 athletes

A dataset containing repetition maximum testing for 12 athletes using 70, 80, and 90% of 1RM

nRM_testing

Format

A data frame with 36 rows and 6 variables:

Athlete

Name of the athlete; ID

1RM

Maximum weight the athlete can lift correctly for a single rep

Goal %1RM

%1RM we want to use for testing; 70, 80, or 90%

Weight

Weight that is estimated to be lifted, but rounded to closes 2.5

Observed %1RM

Recalculated %1RM after rounding the weight

nRM

Number of maximum repetitions performed