Progression table

Progression table explained in the 'Strength Training Manual' Chapter 6.

RIR_increment(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

RIR_increment_fixed_2(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

RIR_increment_fixed_4(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

perc_drop(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

perc_drop_fixed_5(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

perc_drop_fixed_25(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

generate_progression_table(
  type = c("grinding", "ballistic"),
  volume = c("intensive", "normal", "extensive"),
  reps = 1:12,
  step = seq(-3, 0, 1),
  func_max_perc_1RM = get_max_perc_1RM,
  progression_table = RIR_increment
)

Arguments

reps
Number of reps

step
Progression step. Default is 0. Use negative numbers (i.e., -1, -2)

volume
Character string: 'intensive', 'normal' (Default), or 'extensive'

type
Type of max rep table. Options are grinding (Default) and ballistic.

adjustment
Additional adjustment. Default is 0.

func_max_perc_1RM
What max rep table should be used? Default is get_max_perc_1RM

progression_table
Progression table function to use. Default is RIR_increment

Value

List with two elements: adjustment and perc_1RM

Functions

RIR_increment: RIR Increment progression table. This is the original progression table from the Strength Training Manual

RIR_increment_fixed_2: Fixed 2 RIR Increment progression table. This variant have fixed RIR increment across reps from phases to phases (2RIR) and 2RIR difference between extensive, normal, and intensive schemes

RIR_increment_fixed_4: Fixed 4 RIR Increment progression table. This variant have fixed RIR increment across reps from phases to phases (4RIR) and 4RIR difference between extensive, normal, and intensive schemes

perc_drop: Percent Drop progression table. This is the original progression table from the Strength Training Manual

perc_drop_fixed_5: 5% Fixed Percent Drop progression table. This variant have fixed percent drops across reps from phases to phases (5%) and 5% difference between extensive, normal, and intensive schemes

perc_drop_fixed_25: 2.5% Fixed Percent Drop progression table. This variant have fixed percent drops across reps from phases to phases (2.5%) and 2.5% difference between extensive, normal, and intensive schemes

generate_progression_table: Generates progression tables

Examples

RIR_increment(10, step = seq(-3, 0, 1))

#> $adjustment
#> [1] 8.090909 6.272727 4.454545 2.636364
#> 
#> $perc_1RM
#> [1] 0.6240533 0.6485581 0.6750661 0.7038333
#> 

RIR_increment(10, step = seq(-3, 0, 1), volume = "extensive")

#> $adjustment
#> [1] 10.727273  8.909091  7.090909  5.272727
#> 
#> $perc_1RM
#> [1] 0.5916396 0.6136201 0.6372969 0.6628742
#> 

RIR_increment(5, step = seq(-3, 0, 1), type = "ballistic")

#> $adjustment
#> [1] 7.2 5.4 3.6 1.8
#> 
#> $perc_1RM
#> [1] 0.5517181 0.5907931 0.6358249 0.6882881
#> 

RIR_increment(10, step = seq(-3, 0, 1))

#> $adjustment
#> [1] 8.090909 6.272727 4.454545 2.636364
#> 
#> $perc_1RM
#> [1] 0.6240533 0.6485581 0.6750661 0.7038333
#> 

RIR_increment(10, step = seq(-3, 0, 1), volume = "extensive")

#> $adjustment
#> [1] 10.727273  8.909091  7.090909  5.272727
#> 
#> $perc_1RM
#> [1] 0.5916396 0.6136201 0.6372969 0.6628742
#> 

RIR_increment(5, step = seq(-3, 0, 1), type = "ballistic")

#> $adjustment
#> [1] 7.2 5.4 3.6 1.8
#> 
#> $perc_1RM
#> [1] 0.5517181 0.5907931 0.6358249 0.6882881
#> 

RIR_increment(10, step = seq(-3, 0, 1))

#> $adjustment
#> [1] 8.090909 6.272727 4.454545 2.636364
#> 
#> $perc_1RM
#> [1] 0.6240533 0.6485581 0.6750661 0.7038333
#> 

RIR_increment(10, step = seq(-3, 0, 1), volume = "extensive")

#> $adjustment
#> [1] 10.727273  8.909091  7.090909  5.272727
#> 
#> $perc_1RM
#> [1] 0.5916396 0.6136201 0.6372969 0.6628742
#> 

RIR_increment(5, step = seq(-3, 0, 1), type = "ballistic")

#> $adjustment
#> [1] 7.2 5.4 3.6 1.8
#> 
#> $perc_1RM
#> [1] 0.5517181 0.5907931 0.6358249 0.6882881
#> 

perc_drop(10, step = seq(-3, 0, 1))

#> $adjustment
#> [1] -0.18181818 -0.13636364 -0.09090909 -0.04545455
#> 
#> $perc_1RM
#> [1] 0.5683694 0.6138239 0.6592785 0.7047330
#> 

perc_drop(10, step = seq(-3, 0, 1), volume = "extensive")

#> $adjustment
#> [1] -0.22727273 -0.18181818 -0.13636364 -0.09090909
#> 
#> $perc_1RM
#> [1] 0.5229148 0.5683694 0.6138239 0.6592785
#> 

perc_drop(5, step = seq(-3, 0, 1), type = "ballistic")

#> $adjustment
#> [1] -0.170 -0.125 -0.080 -0.035
#> 
#> $perc_1RM
#> [1] 0.5801875 0.6251875 0.6701875 0.7151875
#> 

perc_drop_fixed_5(10, step = seq(-3, 0, 1))

#> $adjustment
#> [1] -0.20 -0.15 -0.10 -0.05
#> 
#> $perc_1RM
#> [1] 0.5501875 0.6001875 0.6501875 0.7001875
#> 

perc_drop_fixed_5(10, step = seq(-3, 0, 1), volume = "extensive")

#> $adjustment
#> [1] -0.16 -0.11 -0.06 -0.01
#> 
#> $perc_1RM
#> [1] 0.5901875 0.6401875 0.6901875 0.7401875
#> 

perc_drop_fixed_5(5, step = seq(-3, 0, 1), type = "ballistic")

#> $adjustment
#> [1] -0.20 -0.15 -0.10 -0.05
#> 
#> $perc_1RM
#> [1] 0.5501875 0.6001875 0.6501875 0.7001875
#> 

perc_drop_fixed_25(10, step = seq(-3, 0, 1))

#> $adjustment
#> [1] -0.100 -0.075 -0.050 -0.025
#> 
#> $perc_1RM
#> [1] 0.6501875 0.6751875 0.7001875 0.7251875
#> 

perc_drop_fixed_25(10, step = seq(-3, 0, 1), volume = "extensive")

#> $adjustment
#> [1] -0.125 -0.100 -0.075 -0.050
#> 
#> $perc_1RM
#> [1] 0.6251875 0.6501875 0.6751875 0.7001875
#> 

perc_drop_fixed_25(5, step = seq(-3, 0, 1), type = "ballistic")

#> $adjustment
#> [1] -0.100 -0.075 -0.050 -0.025
#> 
#> $perc_1RM
#> [1] 0.6501875 0.6751875 0.7001875 0.7251875
#> 

generate_progression_table()

#>          type    volume reps step adjustment  perc_1RM
#> 1    grinding intensive    1   -3   3.000000 0.8824568
#> 2   ballistic intensive    1   -3   3.000000 0.7896399
#> 3    grinding    normal    1   -3   4.000000 0.8572653
#> 4   ballistic    normal    1   -3   4.000000 0.7501875
#> 5    grinding extensive    1   -3   5.000000 0.8334722
#> 6   ballistic extensive    1   -3   5.000000 0.7144899
#> 7    grinding intensive    2   -3   3.272727 0.8506426
#> 8   ballistic intensive    2   -3   3.600000 0.7283533
#> 9    grinding    normal    2   -3   4.454545 0.8230884
#> 10  ballistic    normal    2   -3   4.800000 0.6882881
#> 11   grinding extensive    2   -3   5.636364 0.7972632
#> 12  ballistic extensive    2   -3   6.000000 0.6524008
#> 13   grinding intensive    3   -3   3.545455 0.8210426
#> 14  ballistic intensive    3   -3   4.200000 0.6758949
#> 15   grinding    normal    3   -3   4.909091 0.7915320
#> 16  ballistic    normal    3   -3   5.600000 0.6358249
#> 17   grinding extensive    3   -3   6.272727 0.7640693
#> 18  ballistic extensive    3   -3   7.000000 0.6002401
#> 19   grinding intensive    4   -3   3.818182 0.7934333
#> 20  ballistic intensive    4   -3   4.800000 0.6304852
#> 21   grinding    normal    4   -3   5.363636 0.7623060
#> 22  ballistic    normal    4   -3   6.400000 0.5907931
#> 23   grinding extensive    4   -3   6.909091 0.7335289
#> 24  ballistic extensive    4   -3   8.000000 0.5558026
#> 25   grinding intensive    5   -3   4.090909 0.7676204
#> 26  ballistic intensive    5   -3   5.400000 0.5907931
#> 27   grinding    normal    5   -3   5.818182 0.7351614
#> 28  ballistic    normal    5   -3   7.200000 0.5517181
#> 29   grinding extensive    5   -3   7.545455 0.7053362
#> 30  ballistic extensive    5   -3   9.000000 0.5174912
#> 31   grinding intensive    6   -3   4.363636 0.7434341
#> 32  ballistic intensive    6   -3   6.000000 0.5558026
#> 33   grinding    normal    6   -3   6.272727 0.7098835
#> 34  ballistic    normal    6   -3   8.000000 0.5174912
#> 35   grinding extensive    6   -3   8.181818 0.6792304
#> 36  ballistic extensive    6   -3  10.000000 0.4841208
#> 37   grinding intensive    7   -3   4.636364 0.7207254
#> 38  ballistic intensive    7   -3   6.600000 0.5247250
#> 39   grinding    normal    7   -3   6.727273 0.6862861
#> 40  ballistic    normal    7   -3   8.800000 0.4872629
#> 41   grinding extensive    7   -3   8.818182 0.6549880
#> 42  ballistic extensive    7   -3  11.000000 0.4547935
#> 43   grinding intensive    8   -3   4.909091 0.6993629
#> 44  ballistic intensive    8   -3   7.200000 0.4969389
#> 45   grinding    normal    8   -3   7.181818 0.6642071
#> 46  ballistic    normal    8   -3   9.600000 0.4603712
#> 47   grinding extensive    8   -3   9.454545 0.6324165
#> 48  ballistic extensive    8   -3  12.000000 0.4288165
#> 49   grinding intensive    9   -3   5.181818 0.6792304
#> 50  ballistic intensive    9   -3   7.800000 0.4719474
#> 51   grinding    normal    9   -3   7.636364 0.6435044
#> 52  ballistic    normal    9   -3  10.400000 0.4362926
#> 53   grinding extensive    9   -3  10.090909 0.6113489
#> 54  ballistic extensive    9   -3  13.000000 0.4056466
#> 55   grinding intensive   10   -3   5.454545 0.6602245
#> 56  ballistic intensive   10   -3   8.400000 0.4493493
#> 57   grinding    normal   10   -3   8.090909 0.6240533
#> 58  ballistic    normal   10   -3  11.200000 0.4146074
#> 59   grinding extensive   10   -3  10.727273 0.5916396
#> 60  ballistic extensive   10   -3  14.000000 0.3848522
#> 61   grinding intensive   11   -3   5.727273 0.6422533
#> 62  ballistic intensive   11   -3   9.000000 0.4288165
#> 63   grinding    normal   11   -3   8.545455 0.6057436
#> 64  ballistic    normal   11   -3  12.000000 0.3949759
#> 65   grinding extensive   11   -3  11.363636 0.5731615
#> 66  ballistic extensive   11   -3  15.000000 0.3660858
#> 67   grinding intensive   12   -3   6.000000 0.6252345
#> 68  ballistic intensive   12   -3   9.600000 0.4100781
#> 69   grinding    normal   12   -3   9.000000 0.5884776
#> 70  ballistic    normal   12   -3  12.800000 0.3771194
#> 71   grinding extensive   12   -3  12.000000 0.5558026
#> 72  ballistic extensive   12   -3  16.000000 0.3490645
#> 73   grinding intensive    1   -2   2.000000 0.9091736
#> 74  ballistic intensive    1   -2   2.000000 0.8334722
#> 75   grinding    normal    1   -2   3.000000 0.8824568
#> 76  ballistic    normal    1   -2   3.000000 0.7896399
#> 77   grinding extensive    1   -2   4.000000 0.8572653
#> 78  ballistic extensive    1   -2   4.000000 0.7501875
#> 79   grinding intensive    2   -2   2.181818 0.8777670
#> 80  ballistic intensive    2   -2   2.400000 0.7733713
#> 81   grinding    normal    2   -2   3.363636 0.8484577
#> 82  ballistic    normal    2   -2   3.600000 0.7283533
#> 83   grinding extensive    2   -2   4.545455 0.8210426
#> 84  ballistic extensive    2   -2   4.800000 0.6882881
#> 85   grinding intensive    3   -2   2.363636 0.8484577
#> 86  ballistic intensive    3   -2   2.800000 0.7213550
#> 87   grinding    normal    3   -2   3.727273 0.8169813
#> 88  ballistic    normal    3   -2   4.200000 0.6758949
#> 89   grinding extensive    3   -2   5.090909 0.7877568
#> 90  ballistic extensive    3   -2   5.600000 0.6358249
#> 91   grinding intensive    4   -2   2.545455 0.8210426
#> 92  ballistic intensive    4   -2   3.200000 0.6758949
#> 93   grinding    normal    4   -2   4.090909 0.7877568
#> 94  ballistic    normal    4   -2   4.800000 0.6304852
#> 95   grinding extensive    4   -2   5.636364 0.7570648
#> 96  ballistic extensive    4   -2   6.400000 0.5907931
#> 97   grinding intensive    5   -2   2.727273 0.7953436
#> 98  ballistic intensive    5   -2   3.600000 0.6358249
#> 99   grinding    normal    5   -2   4.454545 0.7605509
#> 100 ballistic    normal    5   -2   5.400000 0.5907931
#> 101  grinding extensive    5   -2   6.181818 0.7286747
#> 102 ballistic extensive    5   -2   7.200000 0.5517181
#> 103  grinding intensive    6   -2   2.909091 0.7712046
#> 104 ballistic intensive    6   -2   4.000000 0.6002401
#> 105  grinding    normal    6   -2   4.818182 0.7351614
#> 106 ballistic    normal    6   -2   6.000000 0.5558026
#> 107  grinding extensive    6   -2   6.727273 0.7023369
#> 108 ballistic extensive    6   -2   8.000000 0.5174912
#> 109  grinding intensive    7   -2   3.090909 0.7484877
#> 110 ballistic intensive    7   -2   4.400000 0.5684273
#> 111  grinding    normal    7   -2   5.181818 0.7114123
#> 112 ballistic    normal    7   -2   6.600000 0.5247250
#> 113  grinding extensive    7   -2   7.272727 0.6778366
#> 114 ballistic extensive    7   -2   8.800000 0.4872629
#> 115  grinding intensive    8   -2   3.272727 0.7270708
#> 116 ballistic intensive    8   -2   4.800000 0.5398169
#> 117  grinding    normal    8   -2   5.545455 0.6891497
#> 118 ballistic    normal    8   -2   7.200000 0.4969389
#> 119  grinding extensive    8   -2   7.818182 0.6549880
#> 120 ballistic extensive    8   -2   9.600000 0.4603712
#> 121  grinding intensive    9   -2   3.454545 0.7068455
#> 122 ballistic intensive    9   -2   5.200000 0.5139486
#> 123  grinding    normal    9   -2   5.909091 0.6682380
#> 124 ballistic    normal    9   -2   7.800000 0.4719474
#> 125  grinding extensive    9   -2   8.363636 0.6336296
#> 126 ballistic extensive    9   -2  10.400000 0.4362926
#> 127  grinding intensive   10   -2   3.636364 0.6877149
#> 128 ballistic intensive   10   -2   5.600000 0.4904461
#> 129  grinding    normal   10   -2   6.272727 0.6485581
#> 130 ballistic    normal   10   -2   8.400000 0.4493493
#> 131  grinding extensive   10   -2   8.909091 0.6136201
#> 132 ballistic extensive   10   -2  11.200000 0.4146074
#> 133  grinding intensive   11   -2   3.818182 0.6695926
#> 134 ballistic intensive   11   -2   6.000000 0.4689992
#> 135  grinding    normal   11   -2   6.636364 0.6300042
#> 136 ballistic    normal   11   -2   9.000000 0.4288165
#> 137  grinding extensive   11   -2   9.454545 0.5948357
#> 138 ballistic extensive   11   -2  12.000000 0.3949759
#> 139  grinding intensive   12   -2   4.000000 0.6524008
#> 140 ballistic intensive   12   -2   6.400000 0.4493493
#> 141  grinding    normal   12   -2   7.000000 0.6124824
#> 142 ballistic    normal   12   -2   9.600000 0.4100781
#> 143  grinding extensive   12   -2  10.000000 0.5771673
#> 144 ballistic extensive   12   -2  12.800000 0.3771194
#> 145  grinding intensive    1   -1   1.000000 0.9375586
#> 146 ballistic intensive    1   -1   1.000000 0.8824568
#> 147  grinding    normal    1   -1   2.000000 0.9091736
#> 148 ballistic    normal    1   -1   2.000000 0.8334722
#> 149  grinding extensive    1   -1   3.000000 0.8824568
#> 150 ballistic extensive    1   -1   3.000000 0.7896399
#> 151  grinding intensive    2   -1   1.090909 0.9066781
#> 152 ballistic intensive    2   -1   1.200000 0.8243208
#> 153  grinding    normal    2   -1   2.272727 0.8754407
#> 154 ballistic    normal    2   -1   2.400000 0.7733713
#> 155  grinding extensive    2   -1   3.454545 0.8462840
#> 156 ballistic extensive    2   -1   3.600000 0.7283533
#> 157  grinding intensive    3   -1   1.181818 0.8777670
#> 158 ballistic intensive    3   -1   1.400000 0.7733713
#> 159  grinding    normal    3   -1   2.545455 0.8441215
#> 160 ballistic    normal    3   -1   2.800000 0.7213550
#> 161  grinding extensive    3   -1   3.909091 0.8129601
#> 162 ballistic extensive    3   -1   4.200000 0.6758949
#> 163  grinding intensive    4   -1   1.272727 0.8506426
#> 164 ballistic intensive    4   -1   1.600000 0.7283533
#> 165  grinding    normal    4   -1   2.818182 0.8149657
#> 166 ballistic    normal    4   -1   3.200000 0.6758949
#> 167  grinding extensive    4   -1   4.363636 0.7821610
#> 168 ballistic extensive    4   -1   4.800000 0.6304852
#> 169  grinding intensive    5   -1   1.363636 0.8251444
#> 170 ballistic intensive    5   -1   1.800000 0.6882881
#> 171  grinding    normal    5   -1   3.090909 0.7877568
#> 172 ballistic    normal    5   -1   3.600000 0.6358249
#> 173  grinding extensive    5   -1   4.818182 0.7536105
#> 174 ballistic extensive    5   -1   5.400000 0.5907931
#> 175  grinding intensive    6   -1   1.454545 0.8011303
#> 176 ballistic intensive    6   -1   2.000000 0.6524008
#> 177  grinding    normal    6   -1   3.363636 0.7623060
#> 178 ballistic    normal    6   -1   4.000000 0.6002401
#> 179  grinding extensive    6   -1   5.272727 0.7270708
#> 180 ballistic extensive    6   -1   6.000000 0.5558026
#> 181  grinding intensive    7   -1   1.545455 0.7784745
#> 182 ballistic intensive    7   -1   2.200000 0.6200704
#> 183  grinding    normal    7   -1   3.636364 0.7384483
#> 184 ballistic    normal    7   -1   4.400000 0.5684273
#> 185  grinding extensive    7   -1   5.727273 0.7023369
#> 186 ballistic extensive    7   -1   6.600000 0.5247250
#> 187  grinding intensive    8   -1   1.636364 0.7570648
#> 188 ballistic intensive    8   -1   2.400000 0.5907931
#> 189  grinding    normal    8   -1   3.909091 0.7160386
#> 190 ballistic    normal    8   -1   4.800000 0.5398169
#> 191  grinding extensive    8   -1   6.181818 0.6792304
#> 192 ballistic extensive    8   -1   7.200000 0.4969389
#> 193  grinding intensive    9   -1   1.727273 0.7368012
#> 194 ballistic intensive    9   -1   2.600000 0.5641558
#> 195  grinding    normal    9   -1   4.181818 0.6949490
#> 196 ballistic    normal    9   -1   5.200000 0.5139486
#> 197  grinding extensive    9   -1   6.636364 0.6575958
#> 198 ballistic extensive    9   -1   7.800000 0.4719474
#> 199  grinding intensive   10   -1   1.818182 0.7175941
#> 200 ballistic intensive   10   -1   2.800000 0.5398169
#> 201  grinding    normal   10   -1   4.454545 0.6750661
#> 202 ballistic    normal   10   -1   5.600000 0.4904461
#> 203  grinding extensive   10   -1   7.090909 0.6372969
#> 204 ballistic extensive   10   -1   8.400000 0.4493493
#> 205  grinding intensive   11   -1   1.909091 0.6993629
#> 206 ballistic intensive   11   -1   3.000000 0.5174912
#> 207  grinding    normal   11   -1   4.727273 0.6562893
#> 208 ballistic    normal   11   -1   6.000000 0.4689992
#> 209  grinding extensive   11   -1   7.545455 0.6182137
#> 210 ballistic extensive   11   -1   9.000000 0.4288165
#> 211  grinding intensive   12   -1   2.000000 0.6820352
#> 212 ballistic intensive   12   -1   3.200000 0.4969389
#> 213  grinding    normal   12   -1   5.000000 0.6385288
#> 214 ballistic    normal   12   -1   6.400000 0.4493493
#> 215  grinding extensive   12   -1   8.000000 0.6002401
#> 216 ballistic extensive   12   -1   9.600000 0.4100781
#> 217  grinding intensive    1    0   0.000000 0.9677732
#> 218 ballistic intensive    1    0   0.000000 0.9375586
#> 219  grinding    normal    1    0   1.000000 0.9375586
#> 220 ballistic    normal    1    0   1.000000 0.8824568
#> 221  grinding extensive    1    0   2.000000 0.9091736
#> 222 ballistic extensive    1    0   2.000000 0.8334722
#> 223  grinding intensive    2    0   0.000000 0.9375586
#> 224 ballistic intensive    2    0   0.000000 0.8824568
#> 225  grinding    normal    2    0   1.181818 0.9041963
#> 226 ballistic    normal    2    0   1.200000 0.8243208
#> 227  grinding extensive    2    0   2.363636 0.8731267
#> 228 ballistic extensive    2    0   2.400000 0.7733713
#> 229  grinding intensive    3    0   0.000000 0.9091736
#> 230 ballistic intensive    3    0   0.000000 0.8334722
#> 231  grinding    normal    3    0   1.363636 0.8731267
#> 232 ballistic    normal    3    0   1.400000 0.7733713
#> 233  grinding extensive    3    0   2.727273 0.8398293
#> 234 ballistic extensive    3    0   2.800000 0.7213550
#> 235  grinding intensive    4    0   0.000000 0.8824568
#> 236 ballistic intensive    4    0   0.000000 0.7896399
#> 237  grinding    normal    4    0   1.545455 0.8441215
#> 238 ballistic    normal    4    0   1.600000 0.7283533
#> 239  grinding extensive    4    0   3.090909 0.8089782
#> 240 ballistic extensive    4    0   3.200000 0.6758949
#> 241  grinding intensive    5    0   0.000000 0.8572653
#> 242 ballistic intensive    5    0   0.000000 0.7501875
#> 243  grinding    normal    5    0   1.727273 0.8169813
#> 244 ballistic    normal    5    0   1.800000 0.6882881
#> 245  grinding extensive    5    0   3.454545 0.7803134
#> 246 ballistic extensive    5    0   3.600000 0.6358249
#> 247  grinding intensive    6    0   0.000000 0.8334722
#> 248 ballistic intensive    6    0   0.000000 0.7144899
#> 249  grinding    normal    6    0   1.909091 0.7915320
#> 250 ballistic    normal    6    0   2.000000 0.6524008
#> 251  grinding extensive    6    0   3.818182 0.7536105
#> 252 ballistic extensive    6    0   4.000000 0.6002401
#> 253  grinding intensive    7    0   0.000000 0.8109642
#> 254 ballistic intensive    7    0   0.000000 0.6820352
#> 255  grinding    normal    7    0   2.090909 0.7676204
#> 256 ballistic    normal    7    0   2.200000 0.6200704
#> 257  grinding extensive    7    0   4.181818 0.7286747
#> 258 ballistic extensive    7    0   4.400000 0.5684273
#> 259  grinding intensive    8    0   0.000000 0.7896399
#> 260 ballistic intensive    8    0   0.000000 0.6524008
#> 261  grinding    normal    8    0   2.272727 0.7451111
#> 262 ballistic    normal    8    0   2.400000 0.5907931
#> 263  grinding extensive    8    0   4.545455 0.7053362
#> 264 ballistic extensive    8    0   4.800000 0.5398169
#> 265  grinding intensive    9    0   0.000000 0.7694083
#> 266 ballistic intensive    9    0   0.000000 0.6252345
#> 267  grinding    normal    9    0   2.454545 0.7238842
#> 268 ballistic    normal    9    0   2.600000 0.5641558
#> 269  grinding extensive    9    0   4.909091 0.6834463
#> 270 ballistic extensive    9    0   5.200000 0.5139486
#> 271  grinding intensive   10    0   0.000000 0.7501875
#> 272 ballistic intensive   10    0   0.000000 0.6002401
#> 273  grinding    normal   10    0   2.636364 0.7038333
#> 274 ballistic    normal   10    0   2.800000 0.5398169
#> 275  grinding extensive   10    0   5.272727 0.6628742
#> 276 ballistic extensive   10    0   5.600000 0.4904461
#> 277  grinding intensive   11    0   0.000000 0.7319037
#> 278 ballistic intensive   11    0   0.000000 0.5771673
#> 279  grinding    normal   11    0   2.818182 0.6848633
#> 280 ballistic    normal   11    0   3.000000 0.5174912
#> 281  grinding extensive   11    0   5.636364 0.6435044
#> 282 ballistic extensive   11    0   6.000000 0.4689992
#> 283  grinding intensive   12    0   0.000000 0.7144899
#> 284 ballistic intensive   12    0   0.000000 0.5558026
#> 285  grinding    normal   12    0   3.000000 0.6668890
#> 286 ballistic    normal   12    0   3.200000 0.4969389
#> 287  grinding extensive   12    0   6.000000 0.6252345
#> 288 ballistic extensive   12    0   6.400000 0.4493493

generate_progression_table(type = "grinding", volume = "normal")

#>        type volume reps step adjustment  perc_1RM
#> 1  grinding normal    1   -3   4.000000 0.8572653
#> 2  grinding normal    2   -3   4.454545 0.8230884
#> 3  grinding normal    3   -3   4.909091 0.7915320
#> 4  grinding normal    4   -3   5.363636 0.7623060
#> 5  grinding normal    5   -3   5.818182 0.7351614
#> 6  grinding normal    6   -3   6.272727 0.7098835
#> 7  grinding normal    7   -3   6.727273 0.6862861
#> 8  grinding normal    8   -3   7.181818 0.6642071
#> 9  grinding normal    9   -3   7.636364 0.6435044
#> 10 grinding normal   10   -3   8.090909 0.6240533
#> 11 grinding normal   11   -3   8.545455 0.6057436
#> 12 grinding normal   12   -3   9.000000 0.5884776
#> 13 grinding normal    1   -2   3.000000 0.8824568
#> 14 grinding normal    2   -2   3.363636 0.8484577
#> 15 grinding normal    3   -2   3.727273 0.8169813
#> 16 grinding normal    4   -2   4.090909 0.7877568
#> 17 grinding normal    5   -2   4.454545 0.7605509
#> 18 grinding normal    6   -2   4.818182 0.7351614
#> 19 grinding normal    7   -2   5.181818 0.7114123
#> 20 grinding normal    8   -2   5.545455 0.6891497
#> 21 grinding normal    9   -2   5.909091 0.6682380
#> 22 grinding normal   10   -2   6.272727 0.6485581
#> 23 grinding normal   11   -2   6.636364 0.6300042
#> 24 grinding normal   12   -2   7.000000 0.6124824
#> 25 grinding normal    1   -1   2.000000 0.9091736
#> 26 grinding normal    2   -1   2.272727 0.8754407
#> 27 grinding normal    3   -1   2.545455 0.8441215
#> 28 grinding normal    4   -1   2.818182 0.8149657
#> 29 grinding normal    5   -1   3.090909 0.7877568
#> 30 grinding normal    6   -1   3.363636 0.7623060
#> 31 grinding normal    7   -1   3.636364 0.7384483
#> 32 grinding normal    8   -1   3.909091 0.7160386
#> 33 grinding normal    9   -1   4.181818 0.6949490
#> 34 grinding normal   10   -1   4.454545 0.6750661
#> 35 grinding normal   11   -1   4.727273 0.6562893
#> 36 grinding normal   12   -1   5.000000 0.6385288
#> 37 grinding normal    1    0   1.000000 0.9375586
#> 38 grinding normal    2    0   1.181818 0.9041963
#> 39 grinding normal    3    0   1.363636 0.8731267
#> 40 grinding normal    4    0   1.545455 0.8441215
#> 41 grinding normal    5    0   1.727273 0.8169813
#> 42 grinding normal    6    0   1.909091 0.7915320
#> 43 grinding normal    7    0   2.090909 0.7676204
#> 44 grinding normal    8    0   2.272727 0.7451111
#> 45 grinding normal    9    0   2.454545 0.7238842
#> 46 grinding normal   10    0   2.636364 0.7038333
#> 47 grinding normal   11    0   2.818182 0.6848633
#> 48 grinding normal   12    0   3.000000 0.6668890

generate_progression_table(
  reps = 1:5,
  step = seq(-5, 0, 1),
  type = "grinding",
  volume = "normal"
)

#>        type volume reps step adjustment  perc_1RM
#> 1  grinding normal    1   -5   6.000000 0.8109642
#> 2  grinding normal    2   -5   6.636364 0.7766442
#> 3  grinding normal    3   -5   7.272727 0.7451111
#> 4  grinding normal    4   -5   7.909091 0.7160386
#> 5  grinding normal    5   -5   8.545455 0.6891497
#> 6  grinding normal    1   -4   5.000000 0.8334722
#> 7  grinding normal    2   -4   5.545455 0.7991921
#> 8  grinding normal    3   -4   6.090909 0.7676204
#> 9  grinding normal    4   -4   6.636364 0.7384483
#> 10 grinding normal    5   -4   7.181818 0.7114123
#> 11 grinding normal    1   -3   4.000000 0.8572653
#> 12 grinding normal    2   -3   4.454545 0.8230884
#> 13 grinding normal    3   -3   4.909091 0.7915320
#> 14 grinding normal    4   -3   5.363636 0.7623060
#> 15 grinding normal    5   -3   5.818182 0.7351614
#> 16 grinding normal    1   -2   3.000000 0.8824568
#> 17 grinding normal    2   -2   3.363636 0.8484577
#> 18 grinding normal    3   -2   3.727273 0.8169813
#> 19 grinding normal    4   -2   4.090909 0.7877568
#> 20 grinding normal    5   -2   4.454545 0.7605509
#> 21 grinding normal    1   -1   2.000000 0.9091736
#> 22 grinding normal    2   -1   2.272727 0.8754407
#> 23 grinding normal    3   -1   2.545455 0.8441215
#> 24 grinding normal    4   -1   2.818182 0.8149657
#> 25 grinding normal    5   -1   3.090909 0.7877568
#> 26 grinding normal    1    0   1.000000 0.9375586
#> 27 grinding normal    2    0   1.181818 0.9041963
#> 28 grinding normal    3    0   1.363636 0.8731267
#> 29 grinding normal    4    0   1.545455 0.8441215
#> 30 grinding normal    5    0   1.727273 0.8169813