Set and Rep Schemes

Set and Rep Schemes

scheme_generic(
  reps = c(5, 5, 5),
  adjustment = c(0, 0, 0),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(),
  progression_table = RIR_increment,
  progression_table_control = list()
)

scheme_wave(
  reps = c(10, 8, 6, 10, 8, 6),
  adjustment = c(4, 2, 0, 6, 4, 2),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(),
  progression_table = RIR_increment,
  progression_table_control = list()
)

Arguments

reps
Numeric vector indicating reps prescription

adjustment
Numeric vector indicating adjustments. Forwarded to progression_table. If the progression_table is RIR_increment, adjustment will be done using RIR. On the other hand, if perc_drop is used, adjustment will be done using 1RM percentage

vertical_planning
Vertical planning function. Default is vertical_linear

vertical_planning_control
Arguments forwarded to the vertical_planning function

progression_table
Progression table function. Default is RIR_increment

progression_table_control
Arguments forwarded to the progression_table function

Value

Data frame with the following columns: reps, index, step, adjustment, and perc_1RM.

Functions

scheme_generic: Generic set and rep scheme. scheme_generic is called in all other set and rep schemes - only the default parameters differ to make easier and quicker schemes writing and groupings

scheme_wave: Wave set and rep scheme

Examples

scheme_generic()

#>    reps index step adjustment  perc_1RM
#> 1     5     1   -3   5.818182 0.7351614
#> 2     5     1   -3   5.818182 0.7351614
#> 3     5     1   -3   5.818182 0.7351614
#> 4     4     2   -2   4.090909 0.7877568
#> 5     4     2   -2   4.090909 0.7877568
#> 6     4     2   -2   4.090909 0.7877568
#> 7     3     3   -1   2.545455 0.8441215
#> 8     3     3   -1   2.545455 0.8441215
#> 9     3     3   -1   2.545455 0.8441215
#> 10    2     4    0   1.181818 0.9041963
#> 11    2     4    0   1.181818 0.9041963
#> 12    2     4    0   1.181818 0.9041963

# Wave set and rep schemes
--------------------------
scheme_wave()

#>    reps index step adjustment  perc_1RM
#> 1    10     1   -3  12.090909 0.5761606
#> 2     8     1   -3   9.181818 0.6360698
#> 3     6     1   -3   6.272727 0.7098835
#> 4    10     1   -3  14.090909 0.5548690
#> 5     8     1   -3  11.181818 0.6102195
#> 6     6     1   -3   8.272727 0.6778366
#> 7     9     2   -2   9.909091 0.6136201
#> 8     7     2   -2   7.181818 0.6792304
#> 9     5     2   -2   4.454545 0.7605509
#> 10    9     2   -2  11.909091 0.5895278
#> 11    7     2   -2   9.181818 0.6498340
#> 12    5     2   -2   6.454545 0.7238842
#> 13    8     3   -1   7.909091 0.6536919
#> 14    6     3   -1   5.363636 0.7254740
#> 15    4     3   -1   2.818182 0.8149657
#> 16    8     3   -1   9.909091 0.6264201
#> 17    6     3   -1   7.363636 0.6920372
#> 18    4     3   -1   4.818182 0.7730093
#> 19    7     4    0   6.090909 0.6964141
#> 20    5     4    0   3.727273 0.7748225
#> 21    3     4    0   1.363636 0.8731267
#> 22    7     4    0   8.090909 0.6655453
#> 23    5     4    0   5.727273 0.7368012
#> 24    3     4    0   3.363636 0.8251444

scheme_wave(
  reps = c(8, 6, 4, 8, 6, 4),
  vertical_planning = vertical_block,
  progression_table = perc_drop,
  progression_table_control = list(type = "ballistic"))

#>    reps index step adjustment  perc_1RM
#> 1     8     1   -2     3.8375 4.4899008
#> 2     6     1   -2     1.8625 2.5769899
#> 3     4     1   -2    -0.1125 0.6771399
#> 4     8     1   -2     5.8375 6.4899008
#> 5     6     1   -2     3.8625 4.5769899
#> 6     4     1   -2     1.8875 2.6771399
#> 7     8     2   -1     3.8975 4.5499008
#> 8     6     2   -1     1.9125 2.6269899
#> 9     4     2   -1    -0.0725 0.7171399
#> 10    8     2   -1     5.8975 6.5499008
#> 11    6     2   -1     3.9125 4.6269899
#> 12    4     2   -1     1.9275 2.7171399
#> 13    8     3    0     3.9575 4.6099008
#> 14    6     3    0     1.9625 2.6769899
#> 15    4     3    0    -0.0325 0.7571399
#> 16    8     3    0     5.9575 6.6099008
#> 17    6     3    0     3.9625 4.6769899
#> 18    4     3    0     1.9675 2.7571399
#> 19    8     4   -3     3.7775 4.4299008
#> 20    6     4   -3     1.8125 2.5269899
#> 21    4     4   -3    -0.1525 0.6371399
#> 22    8     4   -3     5.7775 6.4299008
#> 23    6     4   -3     3.8125 4.5269899
#> 24    4     4   -3     1.8475 2.6371399

# Adjusted second wave
# and using 3 steps progression
scheme_wave(
  reps = c(8, 6, 4, 8, 6, 4),
  # Adjusting using lower %1RM (perc_drop method used)
  adjustment = c(0, 0, 0, -0.1, -0.1, -0.1),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(reps_change = c(0, -2, -4)),
  progression_table = perc_drop,
  progression_table_control = list(volume = "extensive"))

#>    reps index step  adjustment  perc_1RM
#> 1     8     1   -2 -0.16363636 0.6260036
#> 2     6     1   -2 -0.14545455 0.6880177
#> 3     4     1   -2 -0.12727273 0.7551840
#> 4     8     1   -2 -0.26363636 0.5260036
#> 5     6     1   -2 -0.24545455 0.5880177
#> 6     4     1   -2 -0.22727273 0.6551840
#> 7     6     2   -1 -0.10909091 0.7243813
#> 8     4     2   -1 -0.09545455 0.7870022
#> 9     2     2   -1 -0.08181818 0.8557404
#> 10    6     2   -1 -0.20909091 0.6243813
#> 11    4     2   -1 -0.19545455 0.6870022
#> 12    2     2   -1 -0.18181818 0.7557404
#> 13    4     3    0 -0.06363636 0.8188204
#> 14    2     3    0 -0.05454545 0.8830131
#> 15    0     3    0 -0.04545455 0.9545455
#> 16    4     3    0 -0.16363636 0.7188204
#> 17    2     3    0 -0.15454545 0.7830131
#> 18    0     3    0 -0.14545455 0.8545455

# Adjusted using RIR inc
# This time we adjust first wave as well, first two sets easier
scheme_wave(
  reps = c(8, 6, 4, 8, 6, 4),
  # Adjusting using lower %1RM (perc_drop method used)
  adjustment = c(4, 2, 0, 6, 4, 2),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(reps_change = c(0, -2, -4)),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "extensive"))

#>    reps index step adjustment  perc_1RM
#> 1     8     1   -2  11.818182 0.6024294
#> 2     6     1   -2   8.727273 0.6709526
#> 3     4     1   -2   5.636364 0.7570648
#> 4     8     1   -2  13.818182 0.5791912
#> 5     6     1   -2  10.727273 0.6422533
#> 6     4     1   -2   7.636364 0.7207254
#> 7     6     2   -1   9.272727 0.6628742
#> 8     4     2   -1   6.363636 0.7434341
#> 9     2     2   -1   3.454545 0.8462840
#> 10    6     2   -1  11.272727 0.6348473
#> 11    4     2   -1   8.363636 0.7083612
#> 12    2     2   -1   5.454545 0.8011303
#> 13    4     3    0   7.090909 0.7302856
#> 14    2     3    0   4.363636 0.8251444
#> 15    0     3    0   1.636364 0.9483249
#> 16    4     3    0   9.090909 0.6964141
#> 17    2     3    0   6.363636 0.7821610
#> 18    0     3    0   3.636364 0.8919883