Max-Reps Tables

Family of functions to represent max reps tables

get_max_reps_k(perc_1RM, adjustment = 0, type = "grinding", k = 0.0333)

get_max_perc_1RM_k(max_reps, adjustment = 0, type = "grinding", k = 0.0333)

get_predicted_1RM_k(
  weight,
  reps,
  adjustment = 0,
  type = "grinding",
  k = 0.0333
)

get_max_reps_kmod(perc_1RM, adjustment = 0, type = "grinding", kmod = 0.0353)

get_max_perc_1RM_kmod(
  max_reps,
  adjustment = 0,
  type = "grinding",
  kmod = 0.0353
)

get_predicted_1RM_kmod(
  weight,
  reps,
  adjustment = 0,
  type = "grinding",
  kmod = 0.0353
)

get_max_reps_percMR(perc_1RM, adjustment = 1, type = "grinding")

get_max_perc_1RM_percMR(max_reps, adjustment = 1, type = "grinding")

get_predicted_1RM_percMR(weight, reps, adjustment = 1, type = "grinding")

get_max_reps_relInt(perc_1RM, adjustment = 1, type = "grinding")

get_max_perc_1RM_relInt(max_reps, adjustment = 1, type = "grinding")

get_max_reps(perc_1RM, adjustment = 0, type = "grinding")

get_max_perc_1RM(max_reps, adjustment = 0, type = "grinding")

get_predicted_1RM(weight, reps, adjustment = 0, type = "grinding")

Arguments

perc_1RM

Percent of 1RM

adjustment

Adjustment to be implemented, specific to the method

type

Type of max rep table. Options are grinding (Default) and ballistic.

k

User defined k parameter in the Epley's equation. Default is 0.0333

max_reps

Maximum repetitions

weight

Weight used

reps

Number of repetitions done

kmod

User defined kmod parameter in the modified Epley's equation. Default is 0.0353

Functions

get_max_reps_k: Get maximum number of repetitions using the RIR method (adjustment), but with user provided k parameter

get_max_perc_1RM_k: Get maximum %1RM using the RIR method (adjustment), but with user provided k parameter

get_predicted_1RM_k: Get predicted 1RM using the RIR method (adjustment), but with user provided k parameter

get_max_reps_kmod: Get maximum number of repetitions using the RIR method (adjustment), but with user provided kmod parameter using modified Epley's equation

get_max_perc_1RM_kmod: Get maximum %1RM using the RIR method (adjustment), but with user provided kmod parameter using modified Epley's equation

get_predicted_1RM_kmod: Get predicted 1RM using the RIR method (adjustment), but with user provided kmod parameter using modified Epley's equation

get_max_reps_percMR: Get maximum number of repetitions using the %MR (adjustment)

get_max_perc_1RM_percMR: Get maximum %1RM using the %MR (adjustment)

get_predicted_1RM_percMR: Get predicted 1RM using the %MR (adjustment)

get_max_reps_relInt: Get maximum number of repetitions using the Relative Intensity (adjustment)

get_max_perc_1RM_relInt: Get maximum %1RM using the Relative Intensity (adjustment)

get_max_reps: Get maximum number of repetitions using the RIR method (adjustment)

get_max_perc_1RM: Get maximum %1RM using the RIR method (adjustment)

get_predicted_1RM: Get predicted 1RM using the RIR method (adjustment)

Examples

# Get max reps that can be done with 75% 1RM
get_max_reps_k(0.75, k = 0.05)
#> [1] 6.666667

# Get max reps that can be done with 80% 1RM with 2 reps in reserve
get_max_reps_k(
  perc_1RM = 0.8,
  adjustment = 3,
  k = 0.05
)
#> [1] 2

# Get max reps that can be done with 90% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps_k(
  perc_1RM = 0.9,
  adjustment = 2,
  type = "ballistic",
  k = 0.05
)
#> [1] -0.8888889
# Get max %1RM to be used when doing 5 reps to failure
get_max_perc_1RM_k(5, k = 0.0333)
#> [1] 0.8572653

# Get max %1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM_k(
  max_reps = 3,
  adjustment = 2,
  k = 0.05
)
#> [1] 0.8

# Get max %1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM_k(
  max_reps = 3,
  adjustment = 2,
  type = "ballistic",
  k = 0.05
)
#> [1] 0.6666667
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_k(
  weight = 100,
  reps = 5,
  k = 0.05
)
#> [1] 125

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM_k(
  weight = 120,
  reps = 3,
  adjustment = 2,
  k = 0.05
)
#> [1] 150

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM_k(
  weight = 120,
  reps = 2,
  adjustment = 1,
  type = "ballistic",
  k = 0.05
)
#> [1] 156
# Get max reps that can be done with 75% 1RM
get_max_reps_kmod(0.75, kmod = 0.05)
#> [1] 7.666667

# Get max reps that can be done with 80% 1RM with 2 reps in reserve
get_max_reps_kmod(
  perc_1RM = 0.8,
  adjustment = 3,
  kmod = 0.05
)
#> [1] 3

# Get max reps that can be done with 90% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps_kmod(
  perc_1RM = 0.9,
  adjustment = 2,
  type = "ballistic",
  kmod = 0.05
)
#> [1] 0.1111111
# Get max %1RM to be used when doing 5 reps to failure
get_max_perc_1RM_kmod(5, kmod = 0.0353)
#> [1] 0.8762706

# Get max %1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM_kmod(
  max_reps = 3,
  adjustment = 2,
  kmod = 0.05
)
#> [1] 0.8333333

# Get max %1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM_kmod(
  max_reps = 3,
  adjustment = 2,
  type = "ballistic",
  kmod = 0.05
)
#> [1] 0.7142857
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_kmod(
  weight = 100,
  reps = 5,
  kmod = 0.05
)
#> [1] 120

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM_kmod(
  weight = 120,
  reps = 3,
  adjustment = 2,
  kmod = 0.05
)
#> [1] 144

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM_kmod(
  weight = 120,
  reps = 2,
  adjustment = 1,
  type = "ballistic",
  kmod = 0.05
)
#> [1] 144
# Get max reps that can be done with 75% 1RM
get_max_reps_percMR(0.75)
#> [1] 10.01

# Get max reps that can be done with 80% 1RM with 60% of maximum reps
get_max_reps_percMR(
  perc_1RM = 0.8,
  adjustment = 0.6
)
#> [1] 4.5045

# Get max reps that can be done with 90% 1RM with 60% of maximum reps
# using ballistic table
get_max_reps_percMR(
  perc_1RM = 0.9,
  adjustment = 0.6,
  type = "ballistic"
)
#> [1] 1.001
# Get max %1RM to be used when doing 5 reps to failure
get_max_perc_1RM_percMR(5)
#> [1] 0.8572653

# Get max %1RM to be used when doing 3 reps with with 60% of maximum reps
get_max_perc_1RM_percMR(
  max_reps = 3,
  adjustment = 0.6,
)
#> [1] 0.8572653

# Get max %1RM to be used when doing 2 reps with 60% of maximum reps
# using ballistic table
get_max_perc_1RM_percMR(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic"
)
#> [1] 0.7501875
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_percMR(
  weight = 100,
  reps = 5
)
#> [1] 116.65

# Get predicted 1RM when lifting 120kg for 3 reps with 60% of maximum reps
get_predicted_1RM_percMR(
  weight = 120,
  reps = 3,
  adjustment = 0.6
)
#> [1] 139.98

# Get predicted 1RM when lifting 120kg for 2 reps with 60% of maximum reps
# using ballistic table
get_predicted_1RM_percMR(
  weight = 120,
  reps = 2,
  adjustment = 0.6,
  type = "ballistic"
)
#> [1] 146.64
# Get max reps that can be done with 75% 1RM
get_max_reps_relInt(0.75)
#> [1] 10.01

# Get max reps that can be done with 80% 1RM with 60% of Relative Intensity
get_max_reps_relInt(
  perc_1RM = 0.8,
  adjustment = 0.6
)
#> [1] -7.5075

# Get max reps that can be done with 90% 1RM with 60% of Relative Intensity
# using ballistic table
get_max_reps_relInt(
  perc_1RM = 0.9,
  adjustment = 0.6,
  type = "ballistic"
)
#> [1] -5.005
# Get max %1RM to be used when doing 5 reps to failure
get_max_perc_1RM_relInt(5)
#> [1] 0.8572653

# Get max %1RM to be used when doing 3 reps with with 60% of Relative Intensity
get_max_perc_1RM_relInt(
  max_reps = 3,
  adjustment = 0.6,
)
#> [1] 0.5455041

# Get max %1RM to be used when doing 2 reps with 60% of Relative Intensity
# using ballistic table
get_max_perc_1RM_relInt(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic"
)
#> [1] 0.5000833
# Get max reps that can be done with 75% 1RM
get_max_reps(0.75)
#> [1] 10.01

# Get max reps that can be done with 80% 1RM with 2 reps in reserve
get_max_reps(
  perc_1RM = 0.8,
  adjustment = 3
)
#> [1] 4.5075

# Get max reps that can be done with 90% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps(
  perc_1RM = 0.9,
  adjustment = 2,
  type = "ballistic"
)
#> [1] -0.3316667
# Get max %1RM to be used when doing 5 reps to failure
get_max_perc_1RM(5)
#> [1] 0.8572653

# Get max %1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM(
  max_reps = 3,
  adjustment = 2
)
#> [1] 0.8572653

# Get max %1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM(
  max_reps = 3,
  adjustment = 2,
  type = "ballistic"
)
#> [1] 0.7501875
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM(
  weight = 100,
  reps = 5
)
#> [1] 116.65

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM(
  weight = 120,
  reps = 3,
  adjustment = 2
)
#> [1] 139.98

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM(
  weight = 120,
  reps = 2,
  adjustment = 1,
  type = "ballistic"
)
#> [1] 143.976