Set and Rep Schemes

Set and Rep Schemes

scheme_generic(
  reps = c(5, 5, 5),
  adjustment = c(0, 0, 0),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(),
  progression_table = RIR_increment,
  progression_table_control = list()
)

scheme_wave(
  reps = c(10, 8, 6, 10, 8, 6),
  adjustment = c(4, 2, 0, 6, 4, 2),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "extensive")
)

scheme_plateau(
  reps = c(5, 5, 5, 5),
  vertical_planning = vertical_constant,
  vertical_planning_control = list(),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "extensive")
)

scheme_step(
  reps = c(5, 5, 5, 5),
  adjustment = c(-0.3, -0.2, -0.1, 0),
  vertical_planning = vertical_constant,
  vertical_planning_control = list(),
  progression_table = perc_drop,
  progression_table_control = list(volume = "normal")
)

scheme_step_reverse(
  reps = c(10, 10, 10, 10),
  adjustment = c(0, 3, 6, 9),
  vertical_planning = vertical_constant,
  vertical_planning_control = list(),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "normal")
)

scheme_wave_descending(
  reps = c(6, 8, 10, 6, 8, 10),
  adjustment = c(4, 2, 0, 6, 4, 2),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "extensive")
)

scheme_light_heavy(
  reps = c(6, 3, 6, 3, 6, 3),
  adjustment = c(0, -0.2, 0, -0.2, 0, -0.2),
  vertical_planning = vertical_constant,
  vertical_planning_control = list(),
  progression_table = perc_drop,
  progression_table_control = list(volume = "normal")
)

scheme_pyramid(
  reps = c(12, 10, 8, 8, 10, 12),
  adjustment = 0,
  vertical_planning = vertical_linear,
  vertical_planning_control = list(reps_change = c(0, -2, -4, -6)),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "extensive")
)

scheme_pyramid_reverse(
  reps = c(8, 10, 12, 12, 10, 8),
  adjustment = 0,
  vertical_planning = vertical_linear,
  vertical_planning_control = list(reps_change = c(0, -2, -4, -6)),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "extensive")
)

Arguments

reps
Numeric vector indicating reps prescription

adjustment
Numeric vector indicating adjustments. Forwarded to progression_table. If the progression_table is RIR_increment, adjustment will be done using RIR. On the other hand, if perc_drop is used, adjustment will be done using 1RM percentage

vertical_planning
Vertical planning function. Default is vertical_linear

vertical_planning_control
Arguments forwarded to the vertical_planning function

progression_table
Progression table function. Default is RIR_increment

progression_table_control
Arguments forwarded to the progression_table function

Value

Data frame with the following columns: reps, index, step, adjustment, and perc_1RM.

Functions

scheme_generic: Generic set and rep scheme. scheme_generic is called in all other set and rep schemes - only the default parameters differ to make easier and quicker schemes writing and groupings

scheme_wave: Wave set and rep scheme

scheme_plateau: Plateau set and rep scheme

scheme_step: Step set and rep scheme

scheme_step_reverse: Reverse Step set and rep scheme

scheme_wave_descending: Descending Wave set and rep scheme

scheme_light_heavy: Light-Heavy set and rep scheme

scheme_pyramid: Pyramid set and rep scheme

scheme_pyramid_reverse: Reverse Pyramid set and rep scheme

Examples

scheme_generic()

#>    reps index step adjustment  perc_1RM
#> 1     5     1   -3   5.818182 0.7351614
#> 2     5     1   -3   5.818182 0.7351614
#> 3     5     1   -3   5.818182 0.7351614
#> 4     4     2   -2   4.090909 0.7877568
#> 5     4     2   -2   4.090909 0.7877568
#> 6     4     2   -2   4.090909 0.7877568
#> 7     3     3   -1   2.545455 0.8441215
#> 8     3     3   -1   2.545455 0.8441215
#> 9     3     3   -1   2.545455 0.8441215
#> 10    2     4    0   1.181818 0.9041963
#> 11    2     4    0   1.181818 0.9041963
#> 12    2     4    0   1.181818 0.9041963

# Wave set and rep schemes
--------------------------
scheme_wave()

#>    reps index step adjustment  perc_1RM
#> 1    10     1   -3  14.727273 0.5484205
#> 2     8     1   -3  11.454545 0.6068564
#> 3     6     1   -3   8.181818 0.6792304
#> 4    10     1   -3  16.727273 0.5290954
#> 5     8     1   -3  13.454545 0.5832821
#> 6     6     1   -3  10.181818 0.6498340
#> 7     9     2   -2  12.363636 0.5843138
#> 8     7     2   -2   9.272727 0.6485581
#> 9     5     2   -2   6.181818 0.7286747
#> 10    9     2   -2  14.363636 0.5624268
#> 11    7     2   -2  11.272727 0.6217043
#> 12    5     2   -2   8.181818 0.6949490
#> 13    8     3   -1  10.181818 0.6228766
#> 14    6     3   -1   7.272727 0.6934900
#> 15    4     3   -1   4.363636 0.7821610
#> 16    8     3   -1  12.181818 0.5980666
#> 17    6     3   -1   9.272727 0.6628742
#> 18    4     3   -1   6.363636 0.7434341
#> 19    7     4    0   8.181818 0.6642071
#> 20    5     4    0   5.454545 0.7417647
#> 21    3     4    0   2.727273 0.8398293
#> 22    7     4    0  10.181818 0.6360698
#> 23    5     4    0   7.454545 0.7068455
#> 24    3     4    0   4.727273 0.7953436

scheme_wave(
  reps = c(8, 6, 4, 8, 6, 4),
  vertical_planning = vertical_block,
  progression_table = perc_drop,
  progression_table_control = list(type = "ballistic"))

#>    reps index step adjustment  perc_1RM
#> 1     8     1   -2     3.8375 4.4899008
#> 2     6     1   -2     1.8625 2.5769899
#> 3     4     1   -2    -0.1125 0.6771399
#> 4     8     1   -2     5.8375 6.4899008
#> 5     6     1   -2     3.8625 4.5769899
#> 6     4     1   -2     1.8875 2.6771399
#> 7     8     2   -1     3.8975 4.5499008
#> 8     6     2   -1     1.9125 2.6269899
#> 9     4     2   -1    -0.0725 0.7171399
#> 10    8     2   -1     5.8975 6.5499008
#> 11    6     2   -1     3.9125 4.6269899
#> 12    4     2   -1     1.9275 2.7171399
#> 13    8     3    0     3.9575 4.6099008
#> 14    6     3    0     1.9625 2.6769899
#> 15    4     3    0    -0.0325 0.7571399
#> 16    8     3    0     5.9575 6.6099008
#> 17    6     3    0     3.9625 4.6769899
#> 18    4     3    0     1.9675 2.7571399
#> 19    8     4   -3     3.7775 4.4299008
#> 20    6     4   -3     1.8125 2.5269899
#> 21    4     4   -3    -0.1525 0.6371399
#> 22    8     4   -3     5.7775 6.4299008
#> 23    6     4   -3     3.8125 4.5269899
#> 24    4     4   -3     1.8475 2.6371399

# Adjusted second wave
# and using 3 steps progression
scheme_wave(
  reps = c(8, 6, 4, 8, 6, 4),
  # Adjusting using lower %1RM (Perc_Drop method used)
  adjustment = c(0, 0, 0, -0.1, -0.1, -0.1),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(reps_change = c(0, -2, -4)),
  progression_table = perc_drop,
  progression_table_control = list(volume = "extensive"))

#>    reps index step  adjustment  perc_1RM
#> 1     8     1   -2 -0.16363636 0.6260036
#> 2     6     1   -2 -0.14545455 0.6880177
#> 3     4     1   -2 -0.12727273 0.7551840
#> 4     8     1   -2 -0.26363636 0.5260036
#> 5     6     1   -2 -0.24545455 0.5880177
#> 6     4     1   -2 -0.22727273 0.6551840
#> 7     6     2   -1 -0.10909091 0.7243813
#> 8     4     2   -1 -0.09545455 0.7870022
#> 9     2     2   -1 -0.08181818 0.8557404
#> 10    6     2   -1 -0.20909091 0.6243813
#> 11    4     2   -1 -0.19545455 0.6870022
#> 12    2     2   -1 -0.18181818 0.7557404
#> 13    4     3    0 -0.06363636 0.8188204
#> 14    2     3    0 -0.05454545 0.8830131
#> 15    0     3    0 -0.04545455 0.9545455
#> 16    4     3    0 -0.16363636 0.7188204
#> 17    2     3    0 -0.15454545 0.7830131
#> 18    0     3    0 -0.14545455 0.8545455

# Adjusted using RIR inc
# This time we adjust first wave as well, first two sets easier
scheme_wave(
  reps = c(8, 6, 4, 8, 6, 4),
  # Adjusting using lower %1RM (RIR Increment method used)
  adjustment = c(4, 2, 0, 6, 4, 2),
  vertical_planning = vertical_linear,
  vertical_planning_control = list(reps_change = c(0, -2, -4)),
  progression_table = RIR_increment,
  progression_table_control = list(volume = "extensive"))

#>    reps index step adjustment  perc_1RM
#> 1     8     1   -2  11.818182 0.6024294
#> 2     6     1   -2   8.727273 0.6709526
#> 3     4     1   -2   5.636364 0.7570648
#> 4     8     1   -2  13.818182 0.5791912
#> 5     6     1   -2  10.727273 0.6422533
#> 6     4     1   -2   7.636364 0.7207254
#> 7     6     2   -1   9.272727 0.6628742
#> 8     4     2   -1   6.363636 0.7434341
#> 9     2     2   -1   3.454545 0.8462840
#> 10    6     2   -1  11.272727 0.6348473
#> 11    4     2   -1   8.363636 0.7083612
#> 12    2     2   -1   5.454545 0.8011303
#> 13    4     3    0   7.090909 0.7302856
#> 14    2     3    0   4.363636 0.8251444
#> 15    0     3    0   1.636364 0.9483249
#> 16    4     3    0   9.090909 0.6964141
#> 17    2     3    0   6.363636 0.7821610
#> 18    0     3    0   3.636364 0.8919883

# Plateau set and rep schemes
--------------------------
scheme_wave()

#>    reps index step adjustment  perc_1RM
#> 1    10     1   -3  14.727273 0.5484205
#> 2     8     1   -3  11.454545 0.6068564
#> 3     6     1   -3   8.181818 0.6792304
#> 4    10     1   -3  16.727273 0.5290954
#> 5     8     1   -3  13.454545 0.5832821
#> 6     6     1   -3  10.181818 0.6498340
#> 7     9     2   -2  12.363636 0.5843138
#> 8     7     2   -2   9.272727 0.6485581
#> 9     5     2   -2   6.181818 0.7286747
#> 10    9     2   -2  14.363636 0.5624268
#> 11    7     2   -2  11.272727 0.6217043
#> 12    5     2   -2   8.181818 0.6949490
#> 13    8     3   -1  10.181818 0.6228766
#> 14    6     3   -1   7.272727 0.6934900
#> 15    4     3   -1   4.363636 0.7821610
#> 16    8     3   -1  12.181818 0.5980666
#> 17    6     3   -1   9.272727 0.6628742
#> 18    4     3   -1   6.363636 0.7434341
#> 19    7     4    0   8.181818 0.6642071
#> 20    5     4    0   5.454545 0.7417647
#> 21    3     4    0   2.727273 0.8398293
#> 22    7     4    0  10.181818 0.6360698
#> 23    5     4    0   7.454545 0.7068455
#> 24    3     4    0   4.727273 0.7953436

scheme_plateau(
  reps = c(3, 3, 3),
  progression_table_control = list(type = "ballistic"))

#>    reps index step adjustment  perc_1RM
#> 1     3     1   -3        5.6 0.6358249
#> 2     3     1   -3        5.6 0.6358249
#> 3     3     1   -3        5.6 0.6358249
#> 4     3     2   -2        4.2 0.6758949
#> 5     3     2   -2        4.2 0.6758949
#> 6     3     2   -2        4.2 0.6758949
#> 7     3     3   -1        2.8 0.7213550
#> 8     3     3   -1        2.8 0.7213550
#> 9     3     3   -1        2.8 0.7213550
#> 10    3     4    0        1.4 0.7733713
#> 11    3     4    0        1.4 0.7733713
#> 12    3     4    0        1.4 0.7733713

# Step set and rep schemes
--------------------------
scheme_step()

#>    reps index step  adjustment  perc_1RM
#> 1     5     1   -3 -0.43636364 0.4209017
#> 2     5     1   -3 -0.33636364 0.5209017
#> 3     5     1   -3 -0.23636364 0.6209017
#> 4     5     1   -3 -0.13636364 0.7209017
#> 5     5     2   -2 -0.40227273 0.4549926
#> 6     5     2   -2 -0.30227273 0.5549926
#> 7     5     2   -2 -0.20227273 0.6549926
#> 8     5     2   -2 -0.10227273 0.7549926
#> 9     5     3   -1 -0.36818182 0.4890835
#> 10    5     3   -1 -0.26818182 0.5890835
#> 11    5     3   -1 -0.16818182 0.6890835
#> 12    5     3   -1 -0.06818182 0.7890835
#> 13    5     4    0 -0.33409091 0.5231744
#> 14    5     4    0 -0.23409091 0.6231744
#> 15    5     4    0 -0.13409091 0.7231744
#> 16    5     4    0 -0.03409091 0.8231744

scheme_step(
  reps = c(2, 2, 2),
  adjustment = c(-0.1, -0.05, 0),
  vertical_planning = vertical_linear_reverse,
  progression_table_control = list(type = "ballistic")
)

#>    reps index step adjustment  perc_1RM
#> 1     2     1   -3    -0.2175 0.6649568
#> 2     2     1   -3    -0.1675 0.7149568
#> 3     2     1   -3    -0.1175 0.7649568
#> 4     3     2   -2    -0.2000 0.6334722
#> 5     3     2   -2    -0.1500 0.6834722
#> 6     3     2   -2    -0.1000 0.7334722
#> 7     4     3   -1    -0.1725 0.6171399
#> 8     4     3   -1    -0.1225 0.6671399
#> 9     4     3   -1    -0.0725 0.7171399
#> 10    5     4    0    -0.1350 0.6151875
#> 11    5     4    0    -0.0850 0.6651875
#> 12    5     4    0    -0.0350 0.7151875

# Reverse Step set and rep schemes
--------------------------
scheme_step_reverse()

#>    reps index step adjustment  perc_1RM
#> 1    10     1   -3   8.090909 0.6240533
#> 2    10     1   -3  11.090909 0.5874311
#> 3    10     1   -3  14.090909 0.5548690
#> 4    10     1   -3  17.090909 0.5257272
#> 5    10     2   -2   6.272727 0.6485581
#> 6    10     2   -2   9.272727 0.6090943
#> 7    10     2   -2  12.272727 0.5741577
#> 8    10     2   -2  15.272727 0.5430114
#> 9    10     3   -1   4.454545 0.6750661
#> 10   10     3   -1   7.454545 0.6324165
#> 11   10     3   -1  10.454545 0.5948357
#> 12   10     3   -1  13.454545 0.5614708
#> 13   10     4    0   2.636364 0.7038333
#> 14   10     4    0   5.636364 0.6575958
#> 15   10     4    0   8.636364 0.6170589
#> 16   10     4    0  11.636364 0.5812295

# Descending Wave set and rep schemes
--------------------------
scheme_wave_descending()

#>    reps index step adjustment  perc_1RM
#> 1     6     1   -3  12.181818 0.6228766
#> 2     8     1   -3  11.454545 0.6068564
#> 3    10     1   -3  10.727273 0.5916396
#> 4     6     1   -3  14.181818 0.5980666
#> 5     8     1   -3  13.454545 0.5832821
#> 6    10     1   -3  12.727273 0.5692109
#> 7     5     2   -2  10.181818 0.6642071
#> 8     7     2   -2   9.272727 0.6485581
#> 9     9     2   -2   8.363636 0.6336296
#> 10    5     2   -2  12.181818 0.6360698
#> 11    7     2   -2  11.272727 0.6217043
#> 12    9     2   -2  10.363636 0.6079733
#> 13    4     3   -1   8.363636 0.7083612
#> 14    6     3   -1   7.272727 0.6934900
#> 15    8     3   -1   6.181818 0.6792304
#> 16    4     3   -1  10.363636 0.6764485
#> 17    6     3   -1   9.272727 0.6628742
#> 18    8     3   -1   8.181818 0.6498340
#> 19    3     4    0   6.727273 0.7553337
#> 20    5     4    0   5.454545 0.7417647
#> 21    7     4    0   4.181818 0.7286747
#> 22    3     4    0   8.727273 0.7191564
#> 23    5     4    0   7.454545 0.7068455
#> 24    7     4    0   6.181818 0.6949490

# Light-Heavy set and rep schemes
--------------------------
scheme_light_heavy()

#>    reps index step  adjustment  perc_1RM
#> 1     6     1   -3 -0.14545455 0.6880177
#> 2     3     1   -3 -0.31818182 0.5909917
#> 3     6     1   -3 -0.14545455 0.6880177
#> 4     3     1   -3 -0.31818182 0.5909917
#> 5     6     1   -3 -0.14545455 0.6880177
#> 6     3     1   -3 -0.31818182 0.5909917
#> 7     6     2   -2 -0.10909091 0.7243813
#> 8     3     2   -2 -0.28863636 0.6205372
#> 9     6     2   -2 -0.10909091 0.7243813
#> 10    3     2   -2 -0.28863636 0.6205372
#> 11    6     2   -2 -0.10909091 0.7243813
#> 12    3     2   -2 -0.28863636 0.6205372
#> 13    6     3   -1 -0.07272727 0.7607450
#> 14    3     3   -1 -0.25909091 0.6500827
#> 15    6     3   -1 -0.07272727 0.7607450
#> 16    3     3   -1 -0.25909091 0.6500827
#> 17    6     3   -1 -0.07272727 0.7607450
#> 18    3     3   -1 -0.25909091 0.6500827
#> 19    6     4    0 -0.03636364 0.7971086
#> 20    3     4    0 -0.22954545 0.6796281
#> 21    6     4    0 -0.03636364 0.7971086
#> 22    3     4    0 -0.22954545 0.6796281
#> 23    6     4    0 -0.03636364 0.7971086
#> 24    3     4    0 -0.22954545 0.6796281

# Pyramid set and rep schemes
--------------------------
scheme_pyramid()

#>    reps index step adjustment  perc_1RM
#> 1    12     1   -3  12.000000 0.5558026
#> 2    10     1   -3  10.727273 0.5916396
#> 3     8     1   -3   9.454545 0.6324165
#> 4     8     1   -3   9.454545 0.6324165
#> 5    10     1   -3  10.727273 0.5916396
#> 6    12     1   -3  12.000000 0.5558026
#> 7    10     2   -2   8.909091 0.6136201
#> 8     8     2   -2   7.818182 0.6549880
#> 9     6     2   -2   6.727273 0.7023369
#> 10    6     2   -2   6.727273 0.7023369
#> 11    8     2   -2   7.818182 0.6549880
#> 12   10     2   -2   8.909091 0.6136201
#> 13    8     3   -1   6.181818 0.6792304
#> 14    6     3   -1   5.272727 0.7270708
#> 15    4     3   -1   4.363636 0.7821610
#> 16    4     3   -1   4.363636 0.7821610
#> 17    6     3   -1   5.272727 0.7270708
#> 18    8     3   -1   6.181818 0.6792304
#> 19    6     4    0   3.818182 0.7536105
#> 20    4     4    0   3.090909 0.8089782
#> 21    2     4    0   2.363636 0.8731267
#> 22    2     4    0   2.363636 0.8731267
#> 23    4     4    0   3.090909 0.8089782
#> 24    6     4    0   3.818182 0.7536105

# Reverse Pyramid set and rep schemes
--------------------------
scheme_pyramid_reverse()

#>    reps index step adjustment  perc_1RM
#> 1     8     1   -3   9.454545 0.6324165
#> 2    10     1   -3  10.727273 0.5916396
#> 3    12     1   -3  12.000000 0.5558026
#> 4    12     1   -3  12.000000 0.5558026
#> 5    10     1   -3  10.727273 0.5916396
#> 6     8     1   -3   9.454545 0.6324165
#> 7     6     2   -2   6.727273 0.7023369
#> 8     8     2   -2   7.818182 0.6549880
#> 9    10     2   -2   8.909091 0.6136201
#> 10   10     2   -2   8.909091 0.6136201
#> 11    8     2   -2   7.818182 0.6549880
#> 12    6     2   -2   6.727273 0.7023369
#> 13    4     3   -1   4.363636 0.7821610
#> 14    6     3   -1   5.272727 0.7270708
#> 15    8     3   -1   6.181818 0.6792304
#> 16    8     3   -1   6.181818 0.6792304
#> 17    6     3   -1   5.272727 0.7270708
#> 18    4     3   -1   4.363636 0.7821610
#> 19    2     4    0   2.363636 0.8731267
#> 20    4     4    0   3.090909 0.8089782
#> 21    6     4    0   3.818182 0.7536105
#> 22    6     4    0   3.818182 0.7536105
#> 23    4     4    0   3.090909 0.8089782
#> 24    2     4    0   2.363636 0.8731267